Visual studio fixes.

+++ depends/goom-libs/src/goom/src/draw/goom_draw_to_container.h
@@ -75,7 +75,7 @@ inline auto GoomDrawToContainer::GetLastDrawnColor(const int32_t x, const int32_
   {
     return Pixel::BLACK;
   }
-  return colorsList.colorsArray[colorsList.count - 1];
+  return colorsList.colorsArray[static_cast<size_t>(colorsList.count - 1)];
 }
 
 inline auto GoomDrawToContainer::GetLastDrawnColors(const int32_t x, const int32_t y) const
